refactor: modified company codes generator
The commpany ids/codes generator yields ids over the range of the
minimum and maximum company ids found in the IDS.csv files.
This change is a trial hook for an automatic id generator, so its
effectivess is questionable.
Caveat: ids below the min value and above the max value are not
generated. This should be fixed after effectiveness is ascertained.